requirements.txt
pyaudio
pillow
langchain_groq
langchain_fireworks
langchain_core
langchain_community
faiss-cpu
python-dotenv
streamlit
tk